+++ types/src/api/plugins/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! INX plugin API-related types like responses and DTOs.
+
+pub mod indexer;
+pub mod participation;